more info to debug, fix qhull warning and move util to io_modflow
