+++ hydropandas/io/io_modflow.py
@@ -5,8 +5,8 @@ import warnings
 import numpy as np
 import xarray as xr
 from scipy.interpolate import griddata
+from scipy.spatial import Delaunay
 
-from .. import util
 from ..observation import ModelObs
 
 logger = logging.getLogger(__name__)
@@ -57,7 +57,7 @@ def read_imod_results(
 
     xy = np.array([xmid.ravel(), ymid.ravel()]).T
     uv = obs_collection.loc[:, ("x", "y")].dropna(how="any", axis=0).values
-    vtx, wts = util.interp_weights(xy, uv)
+    vtx, wts = interp_weights(xy, uv)
 
     hm_ts = np.zeros((obs_collection.shape[0], len(mtime)))
 
@@ -75,7 +75,7 @@ def read_imod_results(
 
             logger.info(f"read {fname}")
             ihds, _attrs = imod.idf.read(fname)
-            hm = util.interpolate(ihds, vtx, wts)
+            hm = interpolate(ihds, vtx, wts)
             hm_ts[mask, t] = hm[mask]
 
     mo_list = []
@@ -151,7 +151,7 @@ def read_modflow_results(
 
     uv = obs_collection.loc[:, ("x", "y")].dropna(how="any", axis=0).values
     if method == "linear":
-        vtx, wts = util.interp_weights(xy, uv)
+        vtx, wts = interp_weights(xy, uv)
 
     # get interpolated timeseries from hds_arr
     hm_ts = np.nan * np.ones((obs_collection.shape[0], hds_arr.shape[0]))
@@ -169,7 +169,7 @@ def read_modflow_results(
             ihds = hds_arr[t, m]
             ihds[ihds <= -999.0] = np.nan
             if method == "linear":
-                hm = util.interpolate(ihds, vtx, wts)
+                hm = interpolate(ihds, vtx, wts)
                 hm_ts[mask, t] = hm[mask]
             elif method == "nearest":
                 hm_ts[mask, t] = griddata(xy, ihds.ravel(), uv, method=method)[mask]
@@ -191,3 +191,71 @@ def read_modflow_results(
         mo_list.append(mo)
 
     return mo_list
+
+
+
+def interp_weights(xy, uv, d=2):
+    """Calculate interpolation weights [1]_.
+
+    Parameters
+    ----------
+    xy : np.array
+        array containing x-coordinates in first column and y-coordinates
+        in second column
+    uv : np.array
+        array containing coordinates at which interpolation weights should
+        be calculated, x-data in first column and y-data in second column
+    d : int, optional
+        dimension of data? (the default is 2, which works for 2D data)
+
+    Returns
+    -------
+    vertices: np.array
+        array containing interpolation vertices
+    weights: np.array
+        array containing interpolation weights per point
+
+
+    References
+    ----------
+    .. [1] https://stackoverflow.com/questions/20915502/speedup-scipy-griddata-for-multiple-interpolations-between-two-irregular-grids
+    """
+
+    tri = Delaunay(xy)
+    simplex = tri.find_simplex(uv)
+    vertices = np.take(tri.simplices, simplex, axis=0)
+    temp = np.take(tri.transform, simplex, axis=0)
+    delta = uv - temp[:, d]
+    bary = np.einsum("njk,nk->nj", temp[:, :d, :], delta)
+    return vertices, np.hstack((bary, 1 - bary.sum(axis=1, keepdims=True)))
+
+
+def interpolate(values, vtx, wts, fill_value=np.nan):
+    """Interpolate values at locations defined by vertices and points [2]_, as
+    calculated by interp_weights function.
+
+    Parameters
+    ----------
+    values : np.array
+        array containing values to interpolate
+    vtx : np.array
+        array containing interpolation vertices, see interp_weights()
+    wts : np.array
+        array containing interpolation weights, see interp_weights()
+    fill_value : float
+        fill value for points that have to be extrapolated (e.g. at or
+        beyond edges of the known points)
+
+    Returns
+    -------
+    arr: np.array
+        array containing interpolated values at locations as given by
+        vtx and wts
+
+    References
+    ----------
+    .. [2] https://stackoverflow.com/questions/20915502/speedup-scipy-griddata-for-multiple-interpolations-between-two-irregular-grids
+    """
+    ret = np.einsum("nj,nj->n", np.take(values, vtx), wts)
+    ret[np.any(wts < 0, axis=1)] = fill_value
+    return ret
